% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitro_functions.R
\name{nitro_idw}
\alias{nitro_idw}
\title{Interpolación IDW de nitrógeno}
\usage{
nitro_idw(df, raster_base)
}
\arguments{
\item{df}{data.frame con columnas x, y y nitrogeno_suelo}

\item{raster_base}{raster de referencia (SpatRaster) para la interpolación}
}
\value{
raster interpolado (SpatRaster) con valores de nitrógeno estimados
}
\description{
Realiza interpolación IDW usando coordenadas x, y y valores de nitrogeno_suelo
}
