% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitro_functions.R
\name{nitro_prepare}
\alias{nitro_prepare}
\title{Preparar datos de nitrógeno}
\usage{
nitro_prepare(df)
}
\arguments{
\item{df}{data.frame con columnas x, y y nitrogeno_suelo}
}
\value{
data.frame limpio, con valores de nitrogeno_suelo convertidos, duplicados eliminados y sin NA
}
\description{
Limpia los datos y convierte unidades de cg/kg a g/kg
}
